Urban_Issues/requirements.txt
streamlit>=1.37.0
ultralytics>=8.3.0
pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0